packagegroup-*: make ${PN}-hexagon-dsp-binaries RDEPEND on DSP packages
Historically we have been using RRECOMMENDS for ${PN}-firmware
packagegroup so that selection of the firmware packages can be
hand-tuned (which is extremely useful for firmware initramfs images).

For Hexagon DSP packagegroups we don't need such a hand-tuning
(if the binaries are being pulled in, the image can handle few extra MB
in the rootfs). On the other hand havign RRECOMMENDS here can leave the
image without DSP binaries: if there is a version mismatch between
linux-firmware and hexagon-dsp-binaries, the binaries will be silently
left out of the rootfs.

Change RRECOMMENDS to RDEPENDS in order ot make sure that DSP binaries
are always actually pulled into the rootfs if the image requires
installation of the ${PN}-hexagon-dsp-binaries.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@oss.qualcomm.com>