hexagon-dspso: allow building empty Hexagon packages
If the user didn't provide dspso.bin for repacking, OE will skip
generating (empty) packages. This limits our ability to always DEPEND on
them. Currently it's not an issue, since packagegroups use RRECOMMENDS
to pull in DSP packages. Generate empty packages in order to be able to
fulfill dependencies when we switch to RDEPEND.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@oss.qualcomm.com>